askGib:construction: Working, but need dynamodb
* Session is getting too big...

attr: wip - just checking in

issue: relates to #1
